requirements.txt
fastapi
uvicorn
requests
python-dotenv
psycopg2-binary
sqlalchemy